+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - add quiet logging completion of ajax requests
 - update admin overview to use quiet logging
+- add support for multiple groups in User.is_member
 
 ## [6.20.0] - 2022-05-27
 - refactor and enhance icons app